requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
discord.py==2.3.2
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pip==23.3.1
setuptools==68.2.2
wheel==0.41.2
yarl==1.9.4
python-dotenv==1.0.1

